refactor(dbserver): clean up the db server package
