+++ .gitignore
@@ -1,5 +1,5 @@
 config/dev.js
-fxa-auth-db-server/node_modules
+db-server/node_modules
 node_modules/
 sandbox/
 coverage.html
