+++ bin/server.js
@@ -8,7 +8,7 @@
 require('../lib/newrelic')()
 
 var config = require('../config')
-var dbServer = require('../fxa-auth-db-server')
+var dbServer = require('../db-server')
 var error = dbServer.errors
 var logger = require('../lib/logging')('bin.server')
 var DB = require('../lib/db/mysql')(logger, error)
