+++ db-server/lib/error.js
