--- docs/self-host.docker
@@ -1,18 +1,0 @@
-FROM vladikoff/fxa-slim-image:latest
-
-RUN apt-get update \
-  && DEBIAN_FRONTEND=noninteractive apt-get install -yq \
-    mysql-server \
-  && apt-get clean
-COPY . /home/fxa/fxa-auth-db-mysql
-WORKDIR /home/fxa/fxa-auth-db-mysql
-
-RUN npm install --production \
-  && npm cache clear
-
-RUN echo '{"master": { "user": "root" }, "slave": { "user": "root" } }' > config/prod.json
-
-CMD service mysql start && NODE_ENV=prod npm start
-
-# Expose ports
-EXPOSE 8000
