--- fxa-auth-db-server/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "fxa-auth-db-server",
-  "version": "0.48.0",
-  "description": "Firefox Accounts DB Server",
-  "main": "index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mozilla/fxa-auth-db-myqsl"
-  },
-  "author": "Mozilla (https://mozilla.org/)",
-  "license": "MPL-2.0",
-  "bugs": {
-    "url": "https://github.com/mozilla/fxa-auth-db-mysql/issues"
-  },
-  "homepage": "https://github.com/mozilla/fxa-auth-db-mysql",
-  "dependencies": {
-    "restify": "4.3.0",
-    "restify-safe-json-formatter": "0.2.0"
-  },
-  "keywords": [],
-  "engines": {
-    "node": "0.10.x"
-  }
-}
